PMD hardening Step 4: fix casts and formatting in EntityTransform

+++ entity-meta-generator-lib/src/main/java/com/loudsight/meta/serialization/transform/TemporalEntityTransform.java
@@ -49,6 +49,6 @@ public class TemporalEntityTransform extends EntityTransform<Temporal> {
         var millis = readLong(bytes);
         var nanos = readInt(bytes);
 
-        return LocalDateTime.ofEpochSecond(Double.valueOf((((double) millis))/1000.).longValue(), nanos, ZoneOffset.UTC);
+        return LocalDateTime.ofEpochSecond(Double.valueOf(millis / 1000.).longValue(), nanos, ZoneOffset.UTC);
     }
 }
